move helper function to history.ts
